added functionality to visit post author profile

+++ manage_user.py
@@ -1,5 +1,5 @@
 from flask import Flask, redirect, url_for, flash, send_from_directory
-from flask_login import current_user
+from flask_login import current_user, login_required
 from forms import UserDetails
 import os
 from db import User, BlogPost, db
@@ -7,9 +7,10 @@ from functions import img_to_uuid, delete_file
 from flask import Blueprint
 from flask import render_template
 from datetime import date
+from log_reg import logout_user
 
 app = Flask(__name__)
-UPLOAD_USER_IMG = "dynamic/uploads/profile_pic"
+UPLOAD_USER_IMG = "dynamic/profile_pic"
 app.config['UPLOAD_USER_IMG'] = UPLOAD_USER_IMG
 
 user_bp = Blueprint('user_bp', __name__)
@@ -18,6 +19,7 @@ year = date.today().year
 
 
 @user_bp.route("/profile/<int:user_id>", methods=["GET", "POST"])
+@login_required
 def profile(user_id):
     user = User.query.get_or_404(user_id)
     # profile_img_url = url_for('user_bp.profile_img', filename=current_user.profile)
@@ -27,6 +29,7 @@ def profile(user_id):
 
 
 @user_bp.route("/edit_user/<int:user_id>", methods=["GET", "POST"])
+@login_required
 def edit_user(user_id):
     if current_user.is_authenticated:
         user = User.query.get_or_404(user_id)
@@ -56,25 +59,28 @@ def edit_user(user_id):
                     db.session.commit()
                     if form.profile.data:
                         p_pic.data.save(os.path.join(app.config['UPLOAD_USER_IMG'], p_pic_id))
-                        delete_file(profile_path)
+                        if profile_path:
+                            delete_file(profile_path)
 
                     if form.wall.data:
                         wall.data.save(os.path.join(app.config['UPLOAD_USER_IMG'], wall_id))
-                        delete_file(wall_path)
+                        if wall_path:
+                            delete_file(wall_path)
                     return redirect(url_for("user_bp.profile", user_id=user.id))
 
                 except:
                     flash("Error!  Looks like there was a problem...try again!")
-                    return redirect(url_for('user_bp.edit_user'))
+                    return redirect(url_for('user_bp.edit_user', user_id=user.id))
             else:
                 db.session.commit()
                 flash("User Updated Successfully!")
             return redirect(url_for("user_bp.profile", user_id=user.id))
 
-        return render_template("edit-user.html", year=year, current_user=current_user, form=form)
+        return render_template("edit-user.html", year=year, user=user, current_user=current_user, form=form)
 
 
 @user_bp.route('/profile_img/<filename>')
+
 def profile_img(filename):
     return send_from_directory(app.config['UPLOAD_USER_IMG'], filename)
 
@@ -82,3 +88,23 @@ def profile_img(filename):
 @user_bp.route('/wall_img/<filename>')
 def wall_img(filename):
     return send_from_directory(app.config['UPLOAD_USER_IMG'], filename)
+
+
+
+@user_bp.route("/delete/<int:user_id>", methods=["GET", "POST"])
+@login_required
+def delete_user(user_id):
+    user_to_delete = User.query.get_or_404(user_id)
+    profile_path = os.path.join(UPLOAD_USER_IMG, user_to_delete.profile)
+    wall_path = os.path.join(UPLOAD_USER_IMG, user_to_delete.wall)
+    try:
+        delete_file(profile_path)
+        delete_file(wall_path)
+        db.session.delete(user_to_delete)
+        db.session.commit()
+        flash("User was deleted.")
+        logout_user()
+        return redirect(url_for('get_all_posts'))
+    except:
+        flash("Whoops!! There was a problem deleting that post.")
+        return redirect(url_for('get_all_posts'))
